+++ common/src/main/java/net/spell_engine/particle/ParticleHelper.java
@@ -89,13 +89,13 @@ public class ParticleHelper {
             var particle = (ParticleEffect) Registry.PARTICLE_TYPE.get(id);
             var count = batch.count;
             var dynamicallyOffset = requiresDynamicOffset(batch);
-            var defaultOrigin = origin.add(offset(width, batch.shape, batch.rotation, yaw, pitch));
+            var defaultOrigin = origin.add(offset(width, batch.extent, batch.shape, batch.rotation, yaw, pitch));
             if (batch.count < 1) {
                 count = rng.nextFloat() < batch.count ? 1 : 0;
             }
             for(int i = 0; i < count; ++i) {
                 var direction = direction(batch, yaw, pitch);
-                var particleSpecificOrigin = dynamicallyOffset ? origin.add(offset(width, batch.shape, batch.rotation, yaw, pitch)) : defaultOrigin;
+                var particleSpecificOrigin = dynamicallyOffset ? origin.add(offset(width, batch.extent, batch.shape, batch.rotation, yaw, pitch)) : defaultOrigin;
                 world.addParticle(particle, true,
                         particleSpecificOrigin.x, particleSpecificOrigin.y, particleSpecificOrigin.z,
                         direction.x, direction.y, direction.z);
@@ -127,13 +127,13 @@ public class ParticleHelper {
             var particle = (ParticleEffect) Registry.PARTICLE_TYPE.get(id);
             var count = batch.count;
             var dynamicallyOffset = requiresDynamicOffset(batch);
-            var defaultOrigin = origin.add(offset(width, batch.shape, batch.rotation, yaw, pitch));
+            var defaultOrigin = origin.add(offset(width, batch.extent, batch.shape, batch.rotation, yaw, pitch));
             if (batch.count < 1) {
                 count = rng.nextFloat() < batch.count ? 1 : 0;
             }
             for(int i = 0; i < count; ++i) {
                 var direction = direction(batch, yaw, pitch);
-                var particleSpecificOrigin = dynamicallyOffset ? origin.add(offset(width, batch.shape, batch.rotation, yaw, pitch)) : defaultOrigin;
+                var particleSpecificOrigin = dynamicallyOffset ? origin.add(offset(width, batch.extent, batch.shape, batch.rotation, yaw, pitch)) : defaultOrigin;
                 instructions.add(new SpawnInstruction(particle,
                         particleSpecificOrigin.x, particleSpecificOrigin.y, particleSpecificOrigin.z,
                         direction.x, direction.y, direction.z));
@@ -190,18 +190,20 @@ public class ParticleHelper {
         return false;
     }
 
-    private static Vec3d offset(float width, ParticleBatch.Shape shape, ParticleBatch.Rotation rotation, float yaw, float pitch) {
+    private static Vec3d offset(float width, float extent, ParticleBatch.Shape shape, ParticleBatch.Rotation rotation, float yaw, float pitch) {
         var offset = Vec3d.ZERO;
         switch (shape) {
             case CIRCLE, CONE, SPHERE -> {
                 return offset;
             }
             case PIPE -> {
+                var size = width + extent;
                 var angle = (float) Math.toRadians(rng.nextFloat() * 360F);
-                offset = new Vec3d(width,0,0).rotateY(angle);
+                offset = new Vec3d(size,0,0).rotateY(angle);
             }
             case PILLAR -> {
-                var x = randomInRange(0, width);
+                var size = width + extent;
+                var x = randomInRange(0, size);
                 var angle = (float) Math.toRadians(rng.nextFloat() * 360F);
                 offset = new Vec3d(x,0,0).rotateY(angle);
             }